+++ src/app/store/routes/receipt/store-receipt.component.ts
@@ -5,11 +5,12 @@
  * https://opensource.org/licenses/MIT
  */
 
-import { Component } from '@angular/core';
-import { ActivatedRoute } from '@angular/router';
-import { Observable } from 'rxjs';
-import { ReceiptDetail } from 'src/app/models/ReceiptDetail';
-import { StoreReceiptService } from './store-receipt.service';
+import { Component, Inject } from '@angular/core';
+import { ActivatedRoute, Router } from '@angular/router';
+import { catchError, tap } from 'rxjs/operators';
+import { API_SERVICE_INJECTION_TOKENS } from 'src/app/api/api-service-injection-tokens';
+import { IReceiptPublicApiService } from 'src/app/api/receipt-public-api.iservice';
+import { Receipt } from 'src/app/models/Receipt';
 
 @Component({
   selector: 'app-store-receipt',
@@ -18,23 +19,31 @@ import { StoreReceiptService } from './store-receipt.service';
 })
 export class StoreReceiptComponent {
 
-  loading$: Observable<boolean>;
-  details$: Observable<ReceiptDetail[]>;
-  date$: Observable<string>;
+  loading = true;
+  receipt: Receipt | null;
 
   constructor(
-    private service: StoreReceiptService,
-    private route: ActivatedRoute
+    @Inject(API_SERVICE_INJECTION_TOKENS.receipt) private receiptApiService: IReceiptPublicApiService,
+    private route: ActivatedRoute,
+    private router: Router
   ) {
-    this.loading$ = this.service.loading$.pipe();
-    this.details$ = this.service.details$.pipe();
-    this.date$ = this.service.date$.pipe();
-
     this.loadReceipt();
   }
 
   private loadReceipt() {
     const token = this.route.snapshot.paramMap.get('token');
-    this.service.fetchReceipt(token);
+    if (!token) {
+      this.router.navigateByUrl('/');
+    } else {
+      this.receiptApiService.fetchTransactionReceiptByToken(token).pipe(
+        tap(receipt => {
+          this.receipt = receipt;
+          this.loading = false;
+        }),
+        catchError(() => {
+          return this.router.navigateByUrl('/');
+        })
+      ).subscribe();
+    }
   }
 }